Bare bones module which prints on init and exit
Add a very simple module which defines init and exit functions and
prints messages from them.

Create a makefile with targets to build the module, install the module
and clean which all call Kbuild to actually perform the actions.

Also add a very simple Kbuild file and an Overcommit hook to run
cppcheck against the source file.