Check passed paths exist and are directories
Call kern_path() with a flag of LOOKUP_DIRECTORY at init time for each
paths parameter element to check they exist and are directories.  Store
the looked up path structs in an array for later reference.

Add tests for the module failing to load if a missing directory or a
file path are passed.