POC implementation of file 'move' to recycle dir
Initial POC implementation of rescuing unlinked files by creating a
hardlink to the inode when our kprobe pre-handler is called.  When our
kprobe returns the unlink continues as normal, but as there is (now)
always a directory entry referencing the inode it is no-longer removed.

For this initial implementation, just create the new directory entry
directly in the passed recycle directory and add a test for this most
basic behaviour.

As the kprobes feature doesn't provide a built-in way of accessing the
probed function arguments and this varies by platform, only implement
this for x86_64 for now, other platforms can be added as required.