Prevent unlink if an error occurs during recycle
Instead of failing quietly if the recycle function encounters an error,
change the return value of security_path_unlink so that the unlink is
blocked and the error is returned to user space.

Add a test for this behaviour with a file present where a child dir of
the recycle dir should be instead (causing kern_path_create to return
-ENOTDIR).

Implement the new behaviour by switching from registering a kprobe for
security_path_unlink to a kretprobe.  Hook up the 'entry_handler' called
before the probed function itself to do the same work as before but if a
negative (e.g. error) value is returned from recycle(), store the return
value as instance data in the kretprobe_instance struct.  This then
allows the new return handler to change the security_path_unlink return
value to this stored value.

As an optimisation if an error isn't encountered, return non-zero from
the entry_handler as this causes the return handler to be skipped
entirely.  Also, if security_path_unlink was already going to return a
non-zero value then don't change it to our stored value.