Add test for behaviour of unlinking very long path
The current implementation relies on being able to pass the full new
path to kern_path_create (the length of which is limited by PATH_MAX).

Add a test to check if the new path will be more than PATH_MAX that
-ENAMETOOLONG is correctly returned).

Also add logging an error to syslog in this situation to aid future
troubleshooting.