Touch the unlinked files as well as moving them
Add a test for the mtime of a file being updated as part of the 'move'
to the recycle dir.  This is useful as it facilitates simple time-based
cleanup of the recycle dir using find.

Implement this (after creating the new link) by locking the inode,
calling notify_change with the necessary attributes and unlocking it.