Fix compile and tests under Ubuntu 24.04
Instead of struct user_namespace*, kernel 6.3 and above require a struct
idmap* to be passed to vfs_mkdir, notify_change and vfs_link (fetched
from mnt_idmap instead of mnt_user_ns).  Select the correct version at
compile time based on the kernel version.

Turn off a couple of compiler warnings about missing prototypes /
declarations which aren't really relevant as we only have a single
compilation unit.

Fix a couple of test failures by correctly casting rpi->data to an int
pointer before dereferencing it.  Not doing this was causing the return
value to change when stored (rpi->data is a char[]) although oddly this
works under 22.04 for some reason.

Extend the Build & test workflow to run on a 24.04 runner as well as
22.04.