Enable -fanalyzer under Actions & fix issue found
Enable -fanalyzer under GitHub Actions via KCFLAGS environment variable.

Resolve the one issue reported by moving the initialisation of
recycleroot above the first call to buf_add_parent so that there aren't
any gotos to cleanup before it has a value set.  It doesn't seem likely
that this scenario could ever be hit as filenames on most filesystems
are limited to 255 characters and the buffer is much larger than that
but it is good to tidy up anyway.