Switch from hooking path check to inode check
Unfortunately, security_path_unlink is not called when a file is deleted
via NFS as this calls vfs_unlink directly.  However, vfs_unlink calls
security_inode_unlink so make the necessary adjustments to intercept
this call instead.

As security_inode_unlink isn't passed a struct path, the relevant
vfsmount is no-longer available.  Instead, check if the recycle root and
file being unlinked both have inodes linked to the same superblock (e.g.
are on the same file system).

vfs_unlink calls security_path_unlink after locking the inode and
vfs_link also locks it.  This means that we have to unlock the inode
before calling vfs_link and then re-lock it after to avoid a hang.