Fix fail when recycle dir not mounted in namespace
If unlink is being called from within a namespace which doesn't have the
recycle dir mounted in the same (or any) location inside it, trying to
create the necessary subdirectories or the new link fail due to the path
not being found.

Instead of building and passing the full path to kern_path_create, only
build as far as the recycle root to get a relative path and implement
the functions kern_vfs_path_create and vfs_path_create which are
(simplified) equivalents to kern_path_create and filename_create but
take a filename relative to a root path.

Sadly vfs_path_parent_lookup only exists (and lookup_one_qstr_excl is
only exported) from kernel 6.4 and above.  As these are both key to the
new implementation, remove Ubuntu 22.04 from the Actions build matrix
and remove the alternative namespace code for older kernels.

Also add a new test for file deletion within a namespace which doesn't
have the recycle dir mounted within it.