Switch from kprobe to ftrace to fix scheduling bug
Unfortunately k(ret)probes are always called with preemption disabled.
This ends up causing a problem as some fs operations need to wait for
the underlying device and while doing so, be able to be preempted by the
scheduler.  When this is attempted, "BUG: scheduling while atomic" is
written to dmesg along with a stack dump.

The issue only appears to manifest when a large tree of files is being
deleted (or the system is otherwise busy) which is why this doesn't
usually appear during the test script run.

After some investigation (fprobe callbacks also happen with preempt
disabled) thankfully ftrace functions turn out to be called without
first disabling preemption.

Setting the security_inode_unlink return value to non-zero ended up
being slightly more difficult as ftrace functions only trigger on
function entry.  To achieve this, use ftrace_regs_set_return_value and
ftrace_override_function_with_return to return a value without calling
security_inode_unlink at all.  This won't prevent an LSM from blocking
unlink (we only bypass when blocking it ourselves) but could cause
unlink events to not be logged on rare occasions when a recycle error
has occurred.

Sadly the function that the ftrace_override_function_with_return macro
resolves to on x86_64 (override_function_with_return) is not exported,
so copy the implementation into our module (adding appropriate copyright
lines for the original authors).

On the plus side, ftrace provides the ftrace_regs_get_argument macro
which removes the need for platform-specific code to fetch arguments
from the appropriate registers.

Also add a test script which is only run under Actions (as it takes a
lot longer to run).  This loads the module, fetches the Linux kernel
source tree, copies it, deletes it and then uses rsync to compare the
copied and deleted trees for differences (there should be none).  This
reliably reproduced the issue before this commit.