Extend appended epoch suffix from seconds to ms
Currently, if a file with the same name is present at the equivalent
path under the recycle directory the current epoch time in seconds is
appended to the filename.  However, this fails for the situation when
this happens more than once in a given second.

Add a test for this scenario and extend the resolution of the appended
time from seconds to milliseconds.  Also adjust the existing suffix test
to correctly check for milliseconds.